pyproject.toml
[project]
name = "my-backend-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = ["pydantic==2.11.7", "pydantic-settings==2.10.1", "fastapi==0.116.1", "python-dotenv==1.1.1", "redis==6.4.0", "tortoise-orm==0.25.1", "aiomysql==0.2.0"]

[project.optional-dependencies]
dev = ["pre-commit==4.2.0", "pytest==8.4.1", "hypothesis==6.136.3", "coverage==7.9.2", "pytest-cov==6.2.1", "uvicorn==0.23.2", "ipython==8.12.0", "pytest-asyncio==0.21.1", "aiosqlite==0.19.0", "httpx==0.25.0", "fakeredis==2.20.0", "aerich==0.7.2", "tomlkit==0.12.0"]
test = ["pytest>=8.4.1", "hypothesis>=6.136.3", "coverage>=7.9.2", "pytest-cov>=6.2.1"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.aerich]
tortoise_orm = "src.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."